requirements.txt
alpaca-trade-api==1.2.3
astroid==2.6.6
autopep8==1.5.7
bpemb==0.3.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cloudpickle==2.0.0
conllu==4.4.1
cycler==0.10.0
Deprecated==1.2.13
dnspython==2.1.0
filelock==3.0.12
flair==0.9
Flask==2.0.1
Flask-Cors==3.0.10
ftfy==6.0.3
future==0.18.2
gdown==3.12.2
gensim==3.8.3
huggingface-hub==0.0.16
hyperopt==0.2.5
idna==3.2
importlib-metadata==3.10.1
inflection==0.5.1
isort==5.9.3
itsdangerous==2.0.1
Janome==0.4.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
konoha==4.6.5
langdetect==1.0.9
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
more-itertools==8.8.0
mpld3==0.3
msgpack==1.0.2
multitasking==0.0.9
networkx==2.6.3
numpy==1.21.1
overrides==3.1.0
packaging==21.0
pandas==1.3.1
Pillow==8.3.2
praw==7.4.0
prawcore==2.3.0
pycodestyle==2.7.0
pylint==2.9.6
pymongo==3.12.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
Quandl==3.6.1
regex==2021.8.28
requests==2.26.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
segtok==1.5.10
sentencepiece==0.1.95
six==1.16.0
smart-open==5.2.1
sqlitedict==1.7.0
tabulate==0.8.9
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
tqdm==4.62.2
transformers==4.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.6
wcwidth==0.2.5
websocket-client==1.2.1
websockets==9.1
Werkzeug==2.0.1
Wikipedia-API==0.5.4
wrapt==1.12.1
yfinance==0.1.63
zipp==3.5.0
